Code Review: Removing workflow and task
