--- src/main/java/uk/gov/hmcts/reform/divorce/orchestration/workflows/generalreferral/GeneralReferralWorkflow.java
@@ -1,36 +1,0 @@
-package uk.gov.hmcts.reform.divorce.orchestration.workflows.generalreferral;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.tuple.ImmutablePair;
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CaseDetails;
-import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.DefaultWorkflow;
-import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.WorkflowException;
-import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.task.Task;
-import uk.gov.hmcts.reform.divorce.orchestration.tasks.generalreferral.GeneralReferralTask;
-
-import java.util.Map;
-
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.AUTH_TOKEN_JSON_KEY;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.CASE_ID_JSON_KEY;
-
-@Component
-@Slf4j
-@RequiredArgsConstructor
-public class GeneralReferralWorkflow extends DefaultWorkflow<Map<String, Object>> {
-
-    private final GeneralReferralTask generalReferralTask;
-
-    public Map<String, Object> run(CaseDetails caseDetails, String auth) throws WorkflowException {
-
-        return this.execute(
-            new Task[] {
-                generalReferralTask
-            },
-            caseDetails.getCaseData(),
-            ImmutablePair.of(CASE_ID_JSON_KEY, caseDetails.getCaseId()),
-            ImmutablePair.of(AUTH_TOKEN_JSON_KEY, auth)
-        );
-    }
-}
